Two remaining test classes in redis.clients.jedis package are moved to redis.clients.jedis.tests package.
